moves batchjobstatus to it's own module
